+++ crates/medea-reactive/src/subscribers_store/progressable/mod.rs
@@ -3,15 +3,24 @@
 pub mod guarded;
 pub mod recheckable_future;
 
-use std::{cell::RefCell, rc::Rc};
+use std::{
+    cell::RefCell,
+    fmt,
+    pin::Pin,
+    rc::Rc,
+    task::{Context, Poll},
+};
 
-use futures::{channel::mpsc, stream::LocalBoxStream};
+use futures::{
+    channel::mpsc, future::LocalBoxFuture, stream::LocalBoxStream, Future,
+    FutureExt,
+};
 
 use crate::{subscribers_store::SubscribersStore, ObservableCell};
 
 pub use self::{
     guarded::{Guard, Guarded},
-    recheckable_future::{RecheckableCounterFuture, RecheckableFutureExt},
+    recheckable_future::RecheckableFutureExt,
 };
 
 /// [`SubscribersStore`] for progressable collections/field.
@@ -70,5 +79,63 @@ where
     #[inline]
     fn wrap(&self, value: T) -> Guarded<T> {
         Guarded::wrap(value, Rc::clone(&self.counter))
+    }
+}
+
+/// [`RecheckableFutureExt`] for [`SubStore::subscribe`].
+pub struct RecheckableCounterFuture {
+    /// Reference to the [`SubStore::counter`].
+    counter: Rc<ObservableCell<u32>>,
+
+    /// Current [`Future`] which will be polled on
+    /// [`RecheckableCounterFuture::poll`].
+    pending_fut: Option<LocalBoxFuture<'static, ()>>,
+}
+
+impl RecheckableFutureExt for RecheckableCounterFuture {
+    /// Returns `true` if [`RecheckableCounterFuture::counter`] is `0`.
+    fn is_done(&self) -> bool {
+        self.counter.get() == 0
+    }
+
+    /// Refreshes [`RecheckableCounterFuture::pending_fut`] with a new
+    /// [`RecheckableCounterFuture::counter`]'s [`ObservableCell::when_eq`]
+    /// [`Future`].
+    fn restart(&mut self) {
+        self.pending_fut = Some(Box::pin(self.counter.when_eq(0).map(|_| ())));
+    }
+}
+
+impl fmt::Debug for RecheckableCounterFuture {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("RecheckableCounterFuture")
+            .field("counter", &self.counter)
+            .finish()
+    }
+}
+
+impl RecheckableCounterFuture {
+    pub(super) fn new(counter: Rc<ObservableCell<u32>>) -> Self {
+        Self {
+            pending_fut: None,
+            counter,
+        }
+    }
+}
+
+impl Future for RecheckableCounterFuture {
+    type Output = ();
+
+    #[allow(clippy::option_if_let_else)]
+    fn poll(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+    ) -> Poll<Self::Output> {
+        if let Some(fut) = self.pending_fut.as_mut() {
+            fut.as_mut().poll(cx)
+        } else {
+            self.restart();
+            self.poll(cx)
+        }
     }
 }
